rename generator to loader
