+++ pkg/loader/standard.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package generator
+package loader
 
 import (
 	"fmt"
@@ -25,7 +25,7 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-type StandardGenerator struct {
+type StandardLoader struct {
 	workloadDir string
 	parser      *parser.Parser
 	index       int
@@ -37,7 +37,7 @@ type queryInfo struct {
 	sql  string
 }
 
-func (g *StandardGenerator) Next() (string, ast.StmtNode) {
+func (g *StandardLoader) Next() (string, ast.StmtNode) {
 	if g.index == 0 {
 		g.init()
 	}
@@ -72,7 +72,7 @@ func (g *StandardGenerator) Next() (string, ast.StmtNode) {
 	return query.name, stmt[0]
 }
 
-func (g *StandardGenerator) init() error {
+func (g *StandardLoader) init() error {
 	dir := fmt.Sprintf("%s/queries", g.workloadDir)
 	err := filepath.Walk(dir,
 		func(path string, info os.FileInfo, err error) error {
@@ -97,6 +97,6 @@ func (g *StandardGenerator) init() error {
 	return nil
 }
 
-func NewStandardGenerator(workloadDir string) Generator {
-	return &StandardGenerator{workloadDir: workloadDir, parser: parser.New()}
+func NewStandardLoader(workloadDir string) QueryLoader {
+	return &StandardLoader{workloadDir: workloadDir, parser: parser.New()}
 }
